Add SMSG_INITIALIZE_FACTIONS for wrath
