+++ wow_world_messages/src/world/shared/mod.rs
@@ -51,6 +51,8 @@ pub mod cmsg_tutorial_clear123;
 pub mod cmsg_tutorial_flag123;
 pub mod cmsg_tutorial_reset123;
 pub mod environmental_damage_type123;
+pub mod faction_flag1_123_3_5;
+pub mod faction_initializer1_123_3_5;
 pub mod inventory_type1_1223;
 pub mod language11_12;
 pub mod language2_4_33;
@@ -76,6 +78,7 @@ pub mod smsg_chat_player_not_found123;
 pub mod smsg_chat_wrong_faction123;
 pub mod smsg_durability_damage_death123;
 pub mod smsg_environmentaldamagelog123;
+pub mod smsg_initialize_factions1_123_3_5;
 pub mod smsg_login_settimespeed12;
 pub mod smsg_notification123;
 pub mod smsg_pong123;
