+++ wow_world_messages/src/world/vanilla/faction_initializer.rs
@@ -1,45 +1,2 @@
-use std::convert::{TryFrom, TryInto};
-use crate::world::vanilla::FactionFlag;
-use std::io::{Write, Read};
-
-#[derive(Debug, Clone, Copy, PartialEq, Eq, Hash, Default)]
-/// Auto generated from the original `wowm` in file [`wow_message_parser/wowm/world/faction/smsg_initialize_factions.wowm:27`](https://github.com/gtker/wow_messages/tree/main/wow_message_parser/wowm/world/faction/smsg_initialize_factions.wowm#L27):
-/// ```text
-/// struct FactionInitializer {
-///     FactionFlag flag;
-///     u32 standing;
-/// }
-/// ```
-pub struct FactionInitializer {
-    pub flag: FactionFlag,
-    pub standing: u32,
-}
-
-impl FactionInitializer {
-    pub(crate) fn write_into_vec(&self, w: &mut Vec<u8>) -> Result<(), std::io::Error> {
-        // flag: FactionFlag
-        w.write_all(&(self.flag.as_int() as u8).to_le_bytes())?;
-
-        // standing: u32
-        w.write_all(&self.standing.to_le_bytes())?;
-
-        Ok(())
-    }
-}
-
-impl FactionInitializer {
-    pub(crate) fn read<R: std::io::Read>(r: &mut R) -> std::result::Result<Self, crate::errors::ParseError> {
-        // flag: FactionFlag
-        let flag = FactionFlag::new(crate::util::read_u8_le(r)?);
-
-        // standing: u32
-        let standing = crate::util::read_u32_le(r)?;
-
-        Ok(Self {
-            flag,
-            standing,
-        })
-    }
-
-}
+pub use crate::shared::faction_initializer1_123_3_5::FactionInitializer;
