+++ wow_world_messages/src/world/wrath/faction_flag.rs
@@ -1,0 +1,2 @@
+pub use crate::shared::faction_flag1_123_3_5::FactionFlag;
+
