+++ wow_world_messages/src/world/wrath/faction_initializer.rs
@@ -1,0 +1,2 @@
+pub use crate::shared::faction_initializer1_123_3_5::FactionInitializer;
+
