+++ wow_world_messages/src/world/wrath/mod.rs
@@ -175,6 +175,10 @@ pub(crate) mod expansion;
 pub use expansion::*;
 pub(crate) mod extra_movement_flags;
 pub use extra_movement_flags::*;
+pub(crate) mod faction_flag;
+pub use faction_flag::*;
+pub(crate) mod faction_initializer;
+pub use faction_initializer::*;
 pub(crate) mod friend_status;
 pub use friend_status::*;
 pub(crate) mod gender;
@@ -378,6 +382,8 @@ pub(crate) mod smsg_environmentaldamagelog;
 pub use smsg_environmentaldamagelog::*;
 pub(crate) mod smsg_feature_system_status;
 pub use smsg_feature_system_status::*;
+pub(crate) mod smsg_initialize_factions;
+pub use smsg_initialize_factions::*;
 pub(crate) mod smsg_login_settimespeed;
 pub use smsg_login_settimespeed::*;
 pub(crate) mod smsg_login_verify_world;
