+++ wow_world_messages/src/world/wrath/opcodes.rs
@@ -1620,6 +1620,7 @@ use crate::world::wrath::SMSG_TRIGGER_CINEMATIC;
 use crate::world::wrath::SMSG_TUTORIAL_FLAGS;
 use crate::world::wrath::SMSG_EMOTE;
 use crate::world::wrath::SMSG_TEXT_EMOTE;
+use crate::world::wrath::SMSG_INITIALIZE_FACTIONS;
 use crate::world::wrath::SMSG_ACTION_BUTTONS;
 use crate::world::wrath::SMSG_AI_REACTION;
 use crate::world::wrath::SMSG_ATTACKSTART;
@@ -1695,6 +1696,7 @@ pub enum ServerOpcodeMessage {
     SMSG_TUTORIAL_FLAGS(SMSG_TUTORIAL_FLAGS),
     SMSG_EMOTE(SMSG_EMOTE),
     SMSG_TEXT_EMOTE(SMSG_TEXT_EMOTE),
+    SMSG_INITIALIZE_FACTIONS(SMSG_INITIALIZE_FACTIONS),
     SMSG_ACTION_BUTTONS(SMSG_ACTION_BUTTONS),
     SMSG_AI_REACTION(SMSG_AI_REACTION),
     SMSG_ATTACKSTART(SMSG_ATTACKSTART),
@@ -1772,6 +1774,7 @@ impl ServerOpcodeMessage {
             0x00FD => Ok(Self::SMSG_TUTORIAL_FLAGS(<SMSG_TUTORIAL_FLAGS as crate::Message>::read_body(&mut r, body_size)?)),
             0x0103 => Ok(Self::SMSG_EMOTE(<SMSG_EMOTE as crate::Message>::read_body(&mut r, body_size)?)),
             0x0105 => Ok(Self::SMSG_TEXT_EMOTE(<SMSG_TEXT_EMOTE as crate::Message>::read_body(&mut r, body_size)?)),
+            0x0122 => Ok(Self::SMSG_INITIALIZE_FACTIONS(<SMSG_INITIALIZE_FACTIONS as crate::Message>::read_body(&mut r, body_size)?)),
             0x0129 => Ok(Self::SMSG_ACTION_BUTTONS(<SMSG_ACTION_BUTTONS as crate::Message>::read_body(&mut r, body_size)?)),
             0x013C => Ok(Self::SMSG_AI_REACTION(<SMSG_AI_REACTION as crate::Message>::read_body(&mut r, body_size)?)),
             0x0143 => Ok(Self::SMSG_ATTACKSTART(<SMSG_ATTACKSTART as crate::Message>::read_body(&mut r, body_size)?)),
@@ -1998,6 +2001,7 @@ impl ServerOpcodeMessage {
             Self::SMSG_TUTORIAL_FLAGS(c) => c.write_encrypted_server(w, e),
             Self::SMSG_EMOTE(c) => c.write_encrypted_server(w, e),
             Self::SMSG_TEXT_EMOTE(c) => c.write_encrypted_server(w, e),
+            Self::SMSG_INITIALIZE_FACTIONS(c) => c.write_encrypted_server(w, e),
             Self::SMSG_ACTION_BUTTONS(c) => c.write_encrypted_server(w, e),
             Self::SMSG_AI_REACTION(c) => c.write_encrypted_server(w, e),
             Self::SMSG_ATTACKSTART(c) => c.write_encrypted_server(w, e),
@@ -2076,6 +2080,7 @@ impl ServerOpcodeMessage {
             Self::SMSG_TUTORIAL_FLAGS(c) => c.write_unencrypted_server(w),
             Self::SMSG_EMOTE(c) => c.write_unencrypted_server(w),
             Self::SMSG_TEXT_EMOTE(c) => c.write_unencrypted_server(w),
+            Self::SMSG_INITIALIZE_FACTIONS(c) => c.write_unencrypted_server(w),
             Self::SMSG_ACTION_BUTTONS(c) => c.write_unencrypted_server(w),
             Self::SMSG_AI_REACTION(c) => c.write_unencrypted_server(w),
             Self::SMSG_ATTACKSTART(c) => c.write_unencrypted_server(w),
@@ -2154,6 +2159,7 @@ impl ServerOpcodeMessage {
             Self::SMSG_TUTORIAL_FLAGS(c) => c.tokio_write_encrypted_server(w, e).await,
             Self::SMSG_EMOTE(c) => c.tokio_write_encrypted_server(w, e).await,
             Self::SMSG_TEXT_EMOTE(c) => c.tokio_write_encrypted_server(w, e).await,
+            Self::SMSG_INITIALIZE_FACTIONS(c) => c.tokio_write_encrypted_server(w, e).await,
             Self::SMSG_ACTION_BUTTONS(c) => c.tokio_write_encrypted_server(w, e).await,
             Self::SMSG_AI_REACTION(c) => c.tokio_write_encrypted_server(w, e).await,
             Self::SMSG_ATTACKSTART(c) => c.tokio_write_encrypted_server(w, e).await,
@@ -2232,6 +2238,7 @@ impl ServerOpcodeMessage {
             Self::SMSG_TUTORIAL_FLAGS(c) => c.tokio_write_unencrypted_server(w).await,
             Self::SMSG_EMOTE(c) => c.tokio_write_unencrypted_server(w).await,
             Self::SMSG_TEXT_EMOTE(c) => c.tokio_write_unencrypted_server(w).await,
+            Self::SMSG_INITIALIZE_FACTIONS(c) => c.tokio_write_unencrypted_server(w).await,
             Self::SMSG_ACTION_BUTTONS(c) => c.tokio_write_unencrypted_server(w).await,
             Self::SMSG_AI_REACTION(c) => c.tokio_write_unencrypted_server(w).await,
             Self::SMSG_ATTACKSTART(c) => c.tokio_write_unencrypted_server(w).await,
@@ -2310,6 +2317,7 @@ impl ServerOpcodeMessage {
             Self::SMSG_TUTORIAL_FLAGS(c) => c.astd_write_encrypted_server(w, e).await,
             Self::SMSG_EMOTE(c) => c.astd_write_encrypted_server(w, e).await,
             Self::SMSG_TEXT_EMOTE(c) => c.astd_write_encrypted_server(w, e).await,
+            Self::SMSG_INITIALIZE_FACTIONS(c) => c.astd_write_encrypted_server(w, e).await,
             Self::SMSG_ACTION_BUTTONS(c) => c.astd_write_encrypted_server(w, e).await,
             Self::SMSG_AI_REACTION(c) => c.astd_write_encrypted_server(w, e).await,
             Self::SMSG_ATTACKSTART(c) => c.astd_write_encrypted_server(w, e).await,
@@ -2388,6 +2396,7 @@ impl ServerOpcodeMessage {
             Self::SMSG_TUTORIAL_FLAGS(c) => c.astd_write_unencrypted_server(w).await,
             Self::SMSG_EMOTE(c) => c.astd_write_unencrypted_server(w).await,
             Self::SMSG_TEXT_EMOTE(c) => c.astd_write_unencrypted_server(w).await,
+            Self::SMSG_INITIALIZE_FACTIONS(c) => c.astd_write_unencrypted_server(w).await,
             Self::SMSG_ACTION_BUTTONS(c) => c.astd_write_unencrypted_server(w).await,
             Self::SMSG_AI_REACTION(c) => c.astd_write_unencrypted_server(w).await,
             Self::SMSG_ATTACKSTART(c) => c.astd_write_unencrypted_server(w).await,
@@ -2468,6 +2477,7 @@ impl std::fmt::Display for ServerOpcodeMessage {
             ServerOpcodeMessage::SMSG_TUTORIAL_FLAGS(_) => "SMSG_TUTORIAL_FLAGS",
             ServerOpcodeMessage::SMSG_EMOTE(_) => "SMSG_EMOTE",
             ServerOpcodeMessage::SMSG_TEXT_EMOTE(_) => "SMSG_TEXT_EMOTE",
+            ServerOpcodeMessage::SMSG_INITIALIZE_FACTIONS(_) => "SMSG_INITIALIZE_FACTIONS",
             ServerOpcodeMessage::SMSG_ACTION_BUTTONS(_) => "SMSG_ACTION_BUTTONS",
             ServerOpcodeMessage::SMSG_AI_REACTION(_) => "SMSG_AI_REACTION",
             ServerOpcodeMessage::SMSG_ATTACKSTART(_) => "SMSG_ATTACKSTART",
@@ -2716,6 +2726,12 @@ impl From<SMSG_EMOTE> for ServerOpcodeMessage {
 impl From<SMSG_TEXT_EMOTE> for ServerOpcodeMessage {
     fn from(c: SMSG_TEXT_EMOTE) -> Self {
         Self::SMSG_TEXT_EMOTE(c)
+    }
+}
+
+impl From<SMSG_INITIALIZE_FACTIONS> for ServerOpcodeMessage {
+    fn from(c: SMSG_INITIALIZE_FACTIONS) -> Self {
+        Self::SMSG_INITIALIZE_FACTIONS(c)
     }
 }
