+++ wow_world_messages/src/world/wrath/smsg_initialize_factions.rs
@@ -1,0 +1,2 @@
+pub use crate::shared::smsg_initialize_factions1_123_3_5::SMSG_INITIALIZE_FACTIONS;
+
