+++ wowm_language/src/docs/factioninitializer.md
@@ -1,6 +1,6 @@
 # FactionInitializer
 
-## Client Version 1.12
+## Client Version 1.12, Client Version 3.3.5
 
 ### Wowm Representation
